SAIL-MANIP: Potential fix for code scanning alert no. 199: Multiplication result converted to larger type

+++ tests/sail-manip/rotate.c
@@ -46,7 +46,7 @@ static sail_status_t create_test_image(unsigned width,
     image->bytes_per_line = sail_bytes_per_line(width, pixel_format);
 
     /* Allocate pixels */
-    const size_t pixels_size = image->height * image->bytes_per_line;
+    const size_t pixels_size = (size_t)image->height * image->bytes_per_line;
     SAIL_TRY(sail_malloc(pixels_size, &image->pixels));
 
     /* Fill with a simple pattern: each pixel value is row * width + col */
